+++ app/serializers/fallout/character_serializer.rb
@@ -5,11 +5,11 @@ module Fallout
     include Deps[cache: 'cache.avatars']
 
     attributes :provider, :id, :name, :created_at, :avatar, :origin, :abilities, :carry_weight, :initiative, :health, :skills,
-               :tag_skills, :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step, :max_abilities
+               :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step, :max_abilities
 
-    delegate :carry_weight, :initiative, :health, to: :decorator
+    delegate :carry_weight, :initiative, :health, :skills, to: :decorator
     delegate :data, to: :object
-    delegate :origin, :abilities, :skills, :tag_skills, :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step,
+    delegate :origin, :abilities, :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step,
              :max_abilities, to: :data
 
     def provider