+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - fallout 2d20 base integration
 - attributes management for Fallout characters
 - rendering companion features for DH characters
+- skills management for Fallout characters
 
 ## [0.4.9] - 2026-02-23
 ### Added